Major reorg
Got rid of old layouts/ folder, combined user/ folder into a single components/ folder.  Subfolders within components/ describe different sections of the application, and actions/reducers are defined alongside the components that use them the most.
Additionally, all folders have an index.js which serves as the public api for that module. Anything that needs to be imported (components, reducers, etc.) can be imported from the folder itself.  (e.g. import {LoginButton} from './user' rather than import LoginButton from './user/login/LoginButton.js')
