+++ src/app/index.js
@@ -1,0 +1,4 @@
+import App from './App'
+import store from './store'
+
+export {App, store}
