+++ src/components/events/checkin/index.js
@@ -1,0 +1,10 @@
+import EventCheckinAttestor from './EventCheckinAttestor'
+
+import { beginCheckin, endCheckin } from './actions'
+import checkinReducer from './reducer'
+
+/** Exported Components */
+export { EventCheckinAttestor }
+
+/** Exported actions/reducers */
+export { beginCheckin, endCheckin, checkinReducer }
