+++ src/components/events/create/index.js
@@ -1,0 +1,10 @@
+import EventCreator from './EventCreator'
+
+import { createEvent } from './actions'
+import createReducer from './reducer'
+
+/** Exported components */
+export { EventCreator }
+
+/** Exported actions/reducers */
+export { createEvent, createReducer }
