+++ src/components/events/dashboard/index.js
@@ -1,0 +1,4 @@
+import EventDashboard from './EventDashboard'
+
+/** Exported components */
+export { EventDashboard }
