+++ src/components/home/index.js
@@ -1,0 +1,4 @@
+import Home from './Home'
+
+/** Exported Components */
+export { Home }
