+++ src/components/index.js
@@ -1,0 +1,9 @@
+import { Home } from './home'
+import { Profile, LoginButton, LogoutButton } from './user'
+import { EventDashboard, EventCreator, EventCheckinAttestor } from './events'
+
+console.log('components/index')
+console.log(LoginButton)
+
+/** All exported top-level components */
+export { Home, Profile, LoginButton, LogoutButton, EventDashboard, EventCreator, EventCheckinAttestor }
