+++ src/index.js
@@ -1,32 +1,25 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { Router, Route, IndexRoute, browserHistory } from 'react-router'
 import { Provider } from 'react-redux'
 import { syncHistoryWithStore } from 'react-router-redux'
+import { Router, Route, IndexRoute, browserHistory } from 'react-router'
+
+import { App, store } from './app'
 import { UserIsAuthenticated } from './util/wrappers.js'
-
-// Layouts
-import App from './App'
-import Home from './layouts/home/Home'
-import Profile from './user/layouts/profile/Profile'
-import Dashboard from './layouts/dashboard/Dashboard'
-import EventCreator from './layouts/create/EventCreator'
-import CheckinAttestor from './layouts/checkin/CheckinAttestor'
-
-// Redux Store
-import store from './store'
+import { Home, Profile, EventDashboard, EventCreator, EventCheckinAttestor } from './components'
 
 const history = syncHistoryWithStore(browserHistory, store)
 
+// TODO: Maybe this goes in app as well ? 
 ReactDOM.render((
     <Provider store={store}>
       <Router history={history}>
         <Route path="/" component={App}>
           <IndexRoute component={Home} />
-          <Route path="dashboard" component={UserIsAuthenticated(Dashboard)} />
+          <Route path="dashboard" component={UserIsAuthenticated(EventDashboard)} />
           <Route path="profile" component={UserIsAuthenticated(Profile)} />
           <Route path="create" component={UserIsAuthenticated(EventCreator)} />
-          <Route path="checkin" component={UserIsAuthenticated(CheckinAttestor)} />
+          <Route path="checkin" component={UserIsAuthenticated(EventCheckinAttestor)} />
         </Route>
       </Router>
     </Provider>
