--- src/user/ui/loginbutton/LoginButtonContainer.js
@@ -1,24 +1,0 @@
-import { connect } from 'react-redux'
-import LoginButton from './LoginButton'
-import { loginUser } from './LoginButtonActions'
-
-const mapStateToProps = (state, ownProps) => {
-  return {}
-}
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    onLoginUserClick: (event) => {
-      event.preventDefault();
-
-      dispatch(loginUser())
-    }
-  }
-}
-
-const LoginButtonContainer = connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(LoginButton)
-
-export default LoginButtonContainer
