--- src/user/userReducer.js
@@ -1,23 +1,0 @@
-const initialState = {
-  data: null,
-}
-
-const userReducer = (state = initialState, action) => {
-  if (action.type === 'USER_LOGGED_IN')
-  {
-    return Object.assign({}, state, {
-      data: action.payload
-    })
-  }
-
-  if (action.type === 'USER_LOGGED_OUT')
-  {
-    return Object.assign({}, state, {
-      data: null
-    })
-  }
-
-  return state
-}
-
-export default userReducer
