--- src/util/uuid.js
@@ -1,12 +1,0 @@
-
-/**
- * A function for generating a UUID using psuedo-random numbers.  
- * Not necessarily universally unique
- * @returns {String} uuid
- */
-export default function uuidv4() {
-  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
-    var r = Math.random() * 16 | 0, v = c === 'x' ? r : (r & 0x3 | 0x8);
-    return v.toString(16);
-  });
-}
